Create LedNumberButtons.cpp
A C++ program for a FPGA board. This program has the LED states on the board reflect the value of a counter based on the push buttons. If the first button is pressed, the counter increments by 1; if the second button is pressed, the counter decrements by 1; if the third button is pressed, the counter should be shifted right one bit; if the fourth button is pressed, the counter should be shifted left one bit; finally, if multiple buttons are pressed, the LEDs should reflect the state of their respective switches.