Create LEDKeyControl
This C++ program prompts the user for an integer from a terminal connected to both a FPGA board and a USB keyboard., The state of the LEDs are modified by the user’s input. When the user enters a number between 0 and 9, only the LED with that number is reflected by the switch, and all the other switches are ignored. If the user enters -1, all the LEDs reflect the state of their switch.